+++ .github/ISSUE_TEMPLATE/02-bug-battle-assistant.yml
@@ -10,8 +10,10 @@ body:
         ## 前置须知
         
         - 请先将代码更新至最新版本尝试是否依然有问题。
+        - 最新版本代码有问题可尝试使用回滚功能看是否依然有问题。
         - 请先阅读最新的[使用前须知](https://one-dragon.org/zzz/zh/quickstart.html#_2-%E4%BD%BF%E7%94%A8%E5%89%8D%E9%A1%BB%E7%9F%A5)，确保设置正确。
         - 请先阅读最新的[功能说明](https://one-dragon.org/zzz/zh/docs/feat_battle_assistant.html)，确保设置正确。
+        - 请使用脚本F11进行截图上传，其他截图方式有概率不受理。
 
   - type: markdown
     attributes:
